+++ Apps/Examples/Examples/Models/Examples.swift
@@ -90,7 +90,10 @@ struct Examples {
                 type: BuildingExtrusionsExample.self),
         Example(title: "Add a sky layer",
                 description: "Add a customizable sky layer to simulate natural lighting with a Terrain layer.",
-                type: SkyLayerExample.self)
+                type: SkyLayerExample.self),
+        Example(title: "Display a 3D model in a model layer",
+                description: "Showcase the usage of a 3D model layer.",
+                type: ModelLayerExample.self)
     ]
 
     // Examples that focus on annotations.